Add Transaction Controller

+++ community/src/main/java/com/lawencon/community/pojo/voucher/PojoVoucherInsertReq.java
@@ -1,11 +1,12 @@
 package com.lawencon.community.pojo.voucher;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class PojoVoucherInsertReq {
 	private String voucherCode;
 	private LocalDateTime expiredDate;
-	private Integer amount;
+	private BigDecimal amount;
 	
 	public String getVoucherCode() {
 		return voucherCode;
@@ -19,10 +20,10 @@ public class PojoVoucherInsertReq {
 	public void setExpiredDate(LocalDateTime expiredDate) {
 		this.expiredDate = expiredDate;
 	}
-	public Integer getAmount() {
+	public BigDecimal getAmount() {
 		return amount;
 	}
-	public void setAmount(Integer amount) {
+	public void setAmount(BigDecimal amount) {
 		this.amount = amount;
 	}
 }
